move `Type` definiion to types.rs
and move serialize.rs a level up
